Added confirmation dialog, show for any delete from admin tables and pages, and admin bulk actions. Fixed automatic transaction rollback to only for 5xx status codes.

+++ frontend/src/app/pages/admin/clients/clients.component.ts
@@ -9,6 +9,8 @@ import { SnackbarService } from '../../../services/snackbar.service'
 import { RouterLink } from '@angular/router'
 import { SpinnerService } from '../../../services/spinner.service'
 import { OidcInfoComponent } from '../../../components/oidc-info/oidc-info.component'
+import { MatDialog } from '@angular/material/dialog'
+import { ConfirmComponent } from '../../../dialogs/confirm/confirm.component'
 
 export type TableColumn<T> = {
   columnDef: keyof T & string
@@ -51,6 +53,7 @@ export class ClientsComponent implements AfterViewInit {
   private adminService = inject(AdminService)
   private snackbarService = inject(SnackbarService)
   private spinnerService = inject(SpinnerService)
+  private dialog = inject(MatDialog)
 
   async ngAfterViewInit() {
     try {
@@ -64,16 +67,30 @@ export class ClientsComponent implements AfterViewInit {
     }
   }
 
-  async delete(client_id: string) {
-    try {
-      this.spinnerService.show()
-      await this.adminService.deleteClient(client_id)
-      this.dataSource.data = this.dataSource.data.filter(c => c.client_id !== client_id)
-      this.snackbarService.show(`Client ${client_id} was deleted.`)
-    } catch (_e) {
-      this.snackbarService.error('Could not delete client.')
-    } finally {
-      this.spinnerService.hide()
-    }
+  delete(client_id: string) {
+    const dialogRef = this.dialog.open(ConfirmComponent, {
+      data: {
+        message: `Are you sure you want to remove client '${client_id}'?`,
+        header: 'Delete',
+      },
+    })
+
+    dialogRef.afterClosed().subscribe(async (result) => {
+      if (!result) {
+        this.snackbarService.error('Client delete cancelled.')
+        return
+      }
+
+      try {
+        this.spinnerService.show()
+        await this.adminService.deleteClient(client_id)
+        this.dataSource.data = this.dataSource.data.filter(c => c.client_id !== client_id)
+        this.snackbarService.message(`Client ${client_id} was deleted.`)
+      } catch (_e) {
+        this.snackbarService.error('Could not delete client.')
+      } finally {
+        this.spinnerService.hide()
+      }
+    })
   }
 }
